Enhance dashboard load tests
