+++ superset-frontend/cypress-base/cypress/utils/urls.ts
@@ -19,3 +19,4 @@
 
 export const DASHBOARD_LIST = '/dashboard/list/';
 export const CHART_LIST = '/chart/list/';
+export const WORLD_HEALTH_DASHBOARD = '/superset/dashboard/world_health/';
